infrastructure/requirements.txt
# AWS CDK Core
aws-cdk-lib>=2.150.0
constructs>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.100.0

# Development utilities
boto3>=1.34.0
boto3-stubs[dynamodb,s3,lambda,sqs,cognito-idp,apigateway]>=1.34.0

# UUID7 support for job IDs
uuid7>=0.1.0

# Type checking
mypy>=1.8.0
